backend/requirements.txt
Django==6.0.4
django-cors-headers==4.9.0
django-ninja==1.6.2
psycopg2-binary==2.9.12
python-dotenv==1.2.2
